% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_occ_fit.R
\docType{class}
\name{auto_occ_fit-class}
\alias{auto_occ_fit-class}
\title{An S4 class to represent the fitted autologistic occupancy model}
\description{
An S4 class to represent the fitted autologistic occupancy model
}
\section{Slots}{

\describe{
\item{\code{fitType}}{The class of the fitted model object.}

\item{\code{call}}{The formula call.}

\item{\code{formula}}{Double right-hand side formula describing covariates of
detection (rho) and occupancy (psi) in that order.}

\item{\code{y}}{A three-dimensional array of species detections. The first dimension
is sites, the second dimension denotes primary sampling periods, and the third
dimension denotes the secondary sampling periods within each primary sampling
period. If the species was detected on a given survey, that element would receive
a 1, otherwise it is 0. If sampling did not occur for a given sampling period, those
elements should be NA.}

\item{\code{estimates}}{A data.frame of the estimated model parameters and their uncertainty.}

\item{\code{AIC}}{the AIC of the fitted model.}

\item{\code{opt}}{the list returned from the \code{\link[stats]{optim}} call.}

\item{\code{negLogLike}}{The negative log likelihood of the data given the parameters.}

\item{\code{nllFun}}{the likelihood function supplied to \code{\link[stats]{optim}}.}

\item{\code{detcovs}}{A list of the detection covariates supplied to \code{auto_occ}.}

\item{\code{occcovs}}{A list of the occupancy covariates supplied to \code{auto_occ}.}

\item{\code{sites_removed}}{An optional vector that denotes which row indices in y had
no data across all sampled seasons.}
}}

