% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{class}
\name{auto_occ_summary-class}
\alias{auto_occ_summary-class}
\title{An S4 class to represent the summary of a fitted autologistic occupancy model}
\description{
An S4 class to represent the summary of a fitted autologistic occupancy model
}
\section{Slots}{

\describe{
\item{\code{call}}{The formula call.}

\item{\code{optim_convergence_code}}{Convergence code from \code{\link[stats]{optim}}}

\item{\code{optim_iterations}}{Number of iterations from \code{\link[stats]{optim}}}

\item{\code{psi}}{The occupancy parameters in a data.frame}

\item{\code{rho}}{The detection parameters in a data.frame}

\item{\code{AIC}}{the AIC of the fitted model.}
}}

