% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare auto_occ models via AIC}
\usage{
compare_models(model_list, add_formula = FALSE, digits = NULL)
}
\arguments{
\item{model_list}{A list of models of class auto_off_fit. If the list is not
named then the models will be given a numeric identifier based on their location
in the list.}

\item{add_formula}{Whether you want the formula of each model added to the
data.frame. Defaults to FALSE.}

\item{digits}{the number of digits to be reported for each numeric column
in the data.frame. Exists as an easy way to generate what could be a 'close
to publication ready' AIC results table for a given analysis. Defaults to NULL.}
}
\description{
Given a (preferabbly named) list of models, calculate AIC,
 delta AIC, and cumulative model weight.
}
