% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\docType{methods}
\name{confint,auto_occ_fit-method}
\alias{confint,auto_occ_fit-method}
\title{Calculate confidence intervals for model parameters.}
\usage{
\S4method{confint}{auto_occ_fit}(object, parm, level = 0.95, type, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}

\item{parm}{a specification of which parameters to be given confidence
intervals. See details for additional details.}

\item{level}{the confidence level required.}

\item{type}{must be either \code{"psi"} for occupancy or \code{"rho"} for detection.}

\item{...}{Additional arguments to be supplied.}
}
\description{
Calculate confidence intervals for model parameters.
}
