% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\docType{methods}
\name{confint.auto_occ_fit}
\alias{confint.auto_occ_fit}
\title{Calculates confidence intervals for model parameters.}
\usage{
\method{confint}{auto_occ_fit}(object, parm, level = 0.95, type, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}

\item{parm}{a specification of which parameters to be given confidence
intervals. See details for additional details.}

\item{level}{the confidence level required.}

\item{type}{must be either \code{"psi"} for occupancy or \code{"rho"} for detection.}

\item{...}{Additional arguments to be supplied. Not currently used.}
}
\value{
A data.frame with 5 columns that contains the parameter estimates,
standard error, lower, and upper confidence limits for a given
logit-linear predictor, which is specified by \code{type}.
}
\description{
Calculates confidence intervals for model parameters.
}
\details{
Largely used to fill in the rest of the estimates data.frame
for objects of class \code{auto_occ_fit}. However, this could be used
to generate confidence intervals of different levels for a given
\code{auto_occ_fit}.

If \code{parm} is not null, it should be a character vector. While the
theta parameter estimate is titled \code{"psi - theta"}, \code{confint}
uses the \code{type} argument to specify the first part of a parameter
name. Thus, to get only \code{theta} back, \code{type} must equal \code{"psi"}
and \code{parm} must be \code{"theta"}. See last example.
}
\examples{
data("opossum_det_hist")
odh <- opossum_det_hist

# only grabbing the first 30 data.points from each season
# for this example
odh <- split(
  odh,
  factor(odh$Season, levels = unique(odh$Season))
)
odh <- do.call(
  "rbind.data.frame",
  lapply(
    odh,
    function(x) head(x,30)
  )
)


# function to generate detection history
opossum_y <- autoOcc::format_y(
  x = odh,
  site_column = "Site",
  time_column = "Season",
  history_columns = "^Week", # start with Week
  report = FALSE
)

# suppressing warnings because of missing data
m1 <- auto_occ(
  ~1~1,
  y = opossum_y
)
# get 80\% confidence intervals for occupancy
psi_80 <- confint(
  m1,
  level = 0.80,
  type = "psi"
)
# and the same for detection
rho_80 <- confint(
  m1,
  level = 0.80,
  type = "rho"
)

# get just the theta estimate from occupancy,
theta_estimate <- confint(
  m1,
  parm = "theta",
  type = "psi"
)


}
