% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_y.R
\name{format_y}
\alias{format_y}
\title{Format detection data into a 3D array}
\usage{
format_y(x, site_column, time_column, history_columns, report = TRUE)
}
\arguments{
\item{x}{A data.frame in mostly long format that should have one columns that denote
sites sampled, one columns that denotes the primary sampling period, and multiple columns
for the detection history (one column per secondary sampling period within a primary sampling
period). See details for more information.}

\item{site_column}{Either a character that is the column name of the site column or a
numeric or integer the represents the site column's location.}

\item{time_column}{Either a character that is the column name of the primary sampling period column or a
numeric or integer the represents the primary sampling period column's location.}

\item{history_columns}{Either a character object of length 1 that contains a regular expression to
identify the columns that house a species detection history or a numeric or an integer vector that
represents the column locations that house a species detection history.}

\item{report}{Some QA/QC information that the function shares about how the data
get ordered within the detection array and also the names of the history_columns
that were identified by either the regex or numeric/integer vector. Defaults to \code{TRUE}.}
}
\description{
Take a data.frame with detection data and convert it to a
three dimensional array to be used in \code{\link{auto_occ}}.
}
\details{
This function is just a quick way to ensure that more standard representations
of a species detection history get converted into what would be accepted by
\code{\link{auto_occ}}. The classes of the associated columns in \code{x} may vary somewhat
 \code{format_y}. Mor specifically:

 \itemize{
   \item{The site column}{Should preferably be either of class \code{character} or
     \code{factor}. If this column is either a \code{numeric} or \code{integer},
     \code{format_y} will provide a warning that it is converting the column
     to a character class with padded zeroes on the left of the numbers to ensure
     correct sorting (e.g., \code{"001"} instead of \code{"1"}).
   }
   \item{The primary sampling period column}{Can either be of class \code{factor},
   \code{character}, \code{numeric}, or \code{integer}. As a \code{factor}, the rows
   of \code{x} will be sorted based on the levels of the primary sampling period column.
   As a \code{character}, then their order of appearance from the top of \code{x} downwards
   will be used to sort them. As such, when this column is a \code{character}, it
   essentially assumes that the rows are already sorted in the way they should be
   for analysis, with the first primary sampling period on top of \code{x}, followed by the
   second primary sampling period, etc. If this column is either a \code{numeric} or
   \code{integer}, then their numeric order will be used to sort the rows of \code{x}.
   }
   \item{The detection history columns}{should either be of class \code{numeric} or
   \code{integer}. Each element of these columns must either be a 1 if a species was
   detected on a given secondary sampling period within a primary sampling period, a 0 if
   the species was not detected, or an NA if no sampling occurred. This function will
   return an error if any element within the detection history columns are not a \code{0},
   \code{1}, or \code{NA}.
   }
 }
}
