% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_estimates.R
\name{model_list_estimates}
\alias{model_list_estimates}
\title{Collect all parameter estimates from a model list}
\usage{
model_list_estimates(model_list)
}
\arguments{
\item{model_list}{A list of auto_occ_fit objects. If list is named, those names will be retained.
See details for additional information.}
}
\description{
Takes a list of auto_occ_fit objects and retrieves all parameter estimates from them.
}
\details{
This function will return a named list object with two elements. The first
element, estimates, is a data.frame with parameter estimates from each model.
The number of rows will be equal to the total number of unique parameters across
models in the model set. The second element, ses, is a data.frame of standard error
estimates from each model. Like the first element, the number of rows are euqal to the
total number of unique parameters, with \code{NA} values present if that specific
parameter is not in a model.

For the \code{model_list}, the data set must be the same across each model or else
this function will fail.
}
\examples{
 data(opossum_det_hist)
 data(opossum_covariates)

 opossum_y <- autoOcc::format_y(
 x = opossum_det_hist,
 site_column = "Site",
 time_column = "Season",
 history_columns = "^Week" # start with Week
 )

 # going down to just 30 sites for this example
 opossum_y <- opossum_y[1:30,,]

 # scale the covariates for analysis
 oc_scaled <- as.data.frame(
   lapply(
     opossum_covariates[1:30,],
     function(x){
       if(is.numeric(x)){
         scale(x)
       }else{
         x
       }
     }
   )
 )
 m1 <- auto_occ(
   ~1~1,
   y = opossum_y,
   det_covs = oc_scaled,
   occ_covs = oc_scaled
 )

 m2 <- auto_occ(
   ~Income~Income,
   y = opossum_y,
   det_covs = oc_scaled,
   occ_covs = oc_scaled
 )
 m3 <- auto_occ(
   ~1~Income,
   y = opossum_y,
   det_covs = oc_scaled,
   occ_covs = oc_scaled
 )

 m4 <- auto_occ(
   ~Income~1,
   y = opossum_y,
   det_covs = oc_scaled,
   occ_covs = oc_scaled
 )

 all_ests <- model_list_estimates(
   list(m1, m2, m3, m4)
 )

 # you can also name the list
 all_ests <- model_list_estimates(
   list(m1 = m1, m2 = m2, m3 = m3, m4 = m4)
 )


}
