% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opossum_det_hist}
\alias{opossum_det_hist}
\title{Virginia opossum detection / non-detection data from Chicago, IL}
\format{
A data frame with 408 rows and 6 variables:
\describe{
  \item{Site}{Abbreviation for the location sampled}
  \item{Season}{Codes for the different primary sampling periods. JA19 = January 2019,
  AP19 = April 2019, JU19 = July 2019, and OC19 = October 2019}
  \item{Week_1}{For a given Site and Season whether the camera was active detected opossum (1),
  active and did not detect opossum (0), or was not active (NA)
  }
  \item{Week_2}{For a given Site and Season whether the camera was active detected opossum (1),
  active and did not detect opossum (0), or was not active (NA)
  }
  \item{Week_3}{For a given Site and Season whether the camera was active detected opossum (1),
  active and did not detect opossum (0), or was not active (NA)
  }
  \item{Week_4}{For a given Site and Season whether the camera was active detected opossum (1),
  active and did not detect opossum (0), or was not active (NA)
  }
  ...
}
}
\source{
\url{https://doi.org/10.1111/gcb.15800}
}
\usage{
opossum_det_hist
}
\description{
A dataset containing weekly detection / non-detection data
of Virginia opossum (Didelphis virginiana) at 102 unique
camera trapping locations during four primary sampling periods
in 2019 throughout the Chicago greater metropolitan area.
Camera traps were deployed for roughly 4 weeks over
deployment in January, April, July, and October. This is a
subset of the data used in a Urban Wildlife Information
Network (UWIN) publication (see source for link to
article). However, the data have been slightly formatted
from the original data so that there are detection weeks (
the UWIN publication used a Binomial distribution for the
detection model not a Bernoulli like \code{\link{auto_occ}})
does and so some formatting was needed).
}
\keyword{datasets}
