% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{SE}
\alias{SE}
\title{Get Standard Error from autologistic occupancy model.}
\usage{
SE(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}

\item{...}{Other arguments to be passed to \code{vcov.auto_occ_fit}.}
}
\description{
Calculates standard error from the variance covariance matrix
of an \code{"auto_occ_fit"} model.
}
