% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_occ_fit.R
\docType{methods}
\name{summary,auto_occ_fit-method}
\alias{summary,auto_occ_fit-method}
\title{Summarizing autologistic occupancy models}
\usage{
\S4method{summary}{auto_occ_fit}(object)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}
}
\value{
\code{summary} returns an object of class \code{"summary.auto_occ_fit"}.
See \code{\linkS4class{summary.auto_occ_fit}} for additional
details.
}
\description{
summary method for class \code{"auto_occ_fit"}
}
