% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_occ_fit.R
\docType{methods}
\name{summary.auto_occ_fit}
\alias{summary.auto_occ_fit}
\title{Summarizing autologistic occupancy models}
\usage{
\method{summary}{auto_occ_fit}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
\code{summary} returns an object of class \code{"auto_occ_summary"}.
See \code{\linkS4class{auto_occ_summary}} for additional
details.
}
\description{
summary method for class \code{"auto_occ_fit"}
}
\examples{
data("opossum_det_hist")
opossum_y <- format_y(
  opossum_det_hist,
  "Site",
  "Season",
  "^Week",
  report = FALSE
)
# suppressing warnings as there are sites with 0 data
m1 <- suppressWarnings(
  auto_occ(~1~1, y = opossum_y)
)
summary(m1)

}
