% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\docType{methods}
\name{vcov,auto_occ_fit-method}
\alias{vcov,auto_occ_fit-method}
\alias{vcov,auto_occ_fit}
\title{Get variance covariance matrix from autologistic occupancy model.}
\usage{
\S4method{vcov}{auto_occ_fit}(object, type, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}

\item{type}{Either \code{"psi"} for occupancy or \code{"rho"} for detection.
If type is not supplied, provide entire variance covariance matrix.}

\item{...}{Additional arguments. Not currently used.}
}
\description{
Given a \code{"auto_occ_fit"} model. Derive the variance /
covariance matrix for either occupancy (\code{"psi}) or detection (\code{"rho}).
}
