% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\docType{methods}
\name{vcov.auto_occ_fit}
\alias{vcov.auto_occ_fit}
\title{Get variance covariance matrix from autologistic occupancy model.}
\usage{
\method{vcov}{auto_occ_fit}(object, type, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}

\item{type}{Either \code{"psi"} for occupancy or \code{"rho"} for detection.
If type is not supplied, provide entire variance covariance matrix.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
A square matrix of the estimated covariances between the parameter estimates
in the linear predictor of the model.
}
\description{
Given a \code{"auto_occ_fit"} model. Derive the variance /
covariance matrix for either occupancy (\code{"psi"}) or detection (\code{"rho"}).
}
\details{
This function is largely used for making model predictions via
\code{\link{predict.auto_occ_fit}}.
}
\examples{

 data("opossum_det_hist")
 data("opossum_covariates")
 odh <- opossum_det_hist
 oc <- opossum_covariates
 # only grabbing the first 30 data.points from each season
 # for this example
 odh <- split(
   odh,
   factor(odh$Season, levels = unique(odh$Season))
 )
 odh <- do.call(
   "rbind.data.frame",
   lapply(
     odh,
     function(x) head(x,30)
   )
 )
 oc <- head(oc,30)

 # function to generate detection history
 opossum_y <- autoOcc::format_y(
   x = odh,
   site_column = "Site",
   time_column = "Season",
   history_columns = "^Week", # start with Week
   report = FALSE
 )


 # scale the covariates (with base R)
 oc_scaled <- as.data.frame(
   lapply(
     oc,
     function(x){
       if(is.numeric(x)){
         scale(x)
       }else{
         x
       }
     }
   )
 )
 # drop site names
 oc_scaled <- oc_scaled[,-1]
 # suppressing warnings because of missing data
 m1 <- suppressWarnings(
   auto_occ(
     ~Impervious~Impervious+Income,
     y = opossum_y,
     det_covs = oc_scaled,
     occ_covs = oc_scaled
   )
 )
 my_vcov <- vcov(m1)
}
