% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_occ.R
\name{auto_occ}
\alias{auto_occ}
\title{Fit an autologistic occupancy model}
\usage{
auto_occ(
  formula,
  y,
  det_covs = NULL,
  occ_covs = NULL,
  method = "BFGS",
  level = 0.95,
  ...
)
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of
detection (rho) and occupancy (psi) in that order.}

\item{y}{A three-dimensional array of species detections. The first dimension
is sites, the second dimension denotes primary sampling periods, and the third
dimension denotes the secondary sampling periods within each primary sampling
period. If the species was detected on a given survey, that element would receive
a 1, otherwise it is 0. If sampling did not occur for a given sampling period, those
elements should be NA.}

\item{det_covs}{A named list of detection covariates. See details for additional
information on how to format this list.}

\item{occ_covs}{Either a data.frame or named list of occupancy covariates. If
there are temporally varying covariates, use a list. If there are no temporally
varying covariates, use a data.frame. See details for additional infomration
for how to format this.}

\item{method}{The optimization method used by \code{\link[stats]{optim}}.}

\item{level}{The confidence interval size to be calculated for model parameters.
Defaults to 0.95.}

\item{...}{additional arguments. Not used yet.}
}
\description{
Given repeat sampling across multiple primary sampling periods,
this model accounts for first-order temporal autocorrelation in species
occupancy patterns at a site with an additional parameter added to the
latent state of the model (\code{theta}). See Details for additional information.
}
\details{
The autologistic formulation of a standard occupancy model is a simplified
version of a dynamic occupancy model that makes inference on species
patterns of occupancy from one time period to the next such that species
presence at a site during one time step may modify the probability a species
occupies that same site in the following time step. This is done through the
inclusion of a single logit-scale autoregressive term into the model: \eqn{\theta}.
For \eqn{t} in \eqn{1 \dots T} and \eqn{i} in \eqn{1 \dots I} sites, let
\eqn{z_{i,t}} be a species latent occupancy state. For T=1, we do not
know if a species occupied a site in a previous time step
(as sampling has not occurred). Thus, to generate the occupancy probability
at the first time step we have two nearly identical logit linear predictors.
Assuming an intercept only model this would be:

\deqn{logit(\psi_{i,t=1}) = \beta_0}
\deqn{logit(\psi_{i,t=1}^{AR(1)}) = \beta_0 + \theta}

 We don't know if
.
If we assume equilibrium at T = 1, then let \eqn{\psi_{i,t=1}} be the
}
