% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_occ.R
\name{auto_occ}
\alias{auto_occ}
\title{Fit an autologistic occupancy model}
\usage{
auto_occ(
  formula,
  y,
  det_covs = NULL,
  occ_covs = NULL,
  method = "BFGS",
  level = 0.95,
  ...
)
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of
detection (rho) and occupancy (psi) in that order.}

\item{y}{A three-dimensional array of species detections. The first dimension
is sites, the second dimension denotes primary sampling periods, and the third
dimension denotes the secondary sampling periods within each primary sampling
period. If the species was detected on a given survey, that element would receive
a 1, otherwise it is 0. If sampling did not occur for a given sampling period, those
elements should be NA.}

\item{det_covs}{A named list of detection covariates. See details for additional
information on how to format this list.}

\item{occ_covs}{Either a data.frame or named list of occupancy covariates. If
there are temporally varying covariates, use a list. If there are no temporally
varying covariates, use a data.frame. See details for additional infomration
for how to format this.}

\item{method}{The optimization method used by \code{\link[stats]{optim}}.}

\item{level}{The confidence interval size to be calculated for model parameters.
Defaults to 0.95.}

\item{...}{additional arguments. Not used yet.}
}
\description{
Given repeat sampling across multiple primary sampling periods,
this model accounts for first-order temporal autocorrelation in species
occupancy patterns at a site with an additional parameter added to the
latent state of the model (\code{theta}). See Details for additional information.
}
\details{
The autologistic formulation of a standard occupancy model is a simplified
version of a dynamic occupancy model that makes inference on species
patterns of occupancy from one time period to the next such that species
presence at a site during one time step may modify the probability a species
occupies that same site in the following time step. This is done through the
inclusion of a single logit-scale autoregressive term into the model: \eqn{\theta}.
For \eqn{t} in \eqn{1 \dots T} and \eqn{i} in \eqn{1 \dots I} sites, let
\eqn{z_{i,t}} be a species latent occupancy state. For T=1, we do not
know if a species occupied a site in a previous time step
(as sampling has not occurred). Thus, to generate the occupancy probability
at the first time step we have two nearly identical logit linear predictors.
Given a vector of parameters (\eqn{\boldsymbol{\beta}};
 i.e., the occupancy intercept and slope terms) and a matrix of
 covariates whose leading column is a vector of 1's (\eqn{\boldsymbol{X}}),
 the probability of occupancy during the first time step is:

\deqn{\LARGE\psi_{i,t=1} = \frac{\mathrm{ilogit}(\boldsymbol{\beta}\boldsymbol{x}_{i})}{
\mathrm{ilogit}(\boldsymbol{\beta}\boldsymbol{x}_{i}) + (1 - \mathrm{ilogit}(\boldsymbol{\beta}\boldsymbol{x}_{i} + \theta)}}

\deqn{\Large z_{i,t=1}\sim \mathrm{Bernoulli}(\psi_{i,t=1})}

Where ilogit is the inverse logit-link. This may seem somewhat complicated for
setting the occupancy of the initial sampling period, but this is the way to
derive the expected occupancy from an autologistic occupancy model. We do this
because we do not know species presence before we start sampling, and so
we condition on the possibility of both states (either the species was there or was not in t-1). Given
this model parameterization, we essentially assume that the population is at equilibrium in the
first sampling period.

Following the first time period, we can model the rest of the latent state as

\deqn{\Large z_{i,t}\sim \mathrm{Bernoulli}(\psi_{i,t}), t>1}

where

\deqn{\Large \mathrm{logit}(\psi{i,t}) = \boldsymbol{\beta}\boldsymbol{x}_{i} + \theta \times z_{i,t-1} }

Note that it is absolutely possible to have time-varying covariates in this
portion of the model, and \code{auto_occ()} can accommodate this.


For \eqn{j} in \eqn{1, \dots, J} repeated samples during each sample period,
the observational or detection model is:

\deqn{\Large y_{i,t,j}|z_{i,t} \sim \mathrm{Bernoulli}(\rho_{i,t,j} \times z_{i,t,j})}

where \eqn{\rho_{i,t,j}} is the probability of detecting the species at site i, primary sampling
period t, and secondary sample j given the species presence. This probability can
be made a function of covariates with the logit link. For ease of explanation,
I am assuming there is only spatial variation in a species detection probability (
although you can incorporate variation across primary and secondary sampling periods
if you wish to do so). For \eqn{d} in \eqn{1, \dots, D} detection parameters which includes
the intercept let \eqn{\boldsymbol{a}} be  a vector of parameters and \eqn{\boldsymbol{W}} be an \eqn{I \times D} design matrix whose first column
is a vector of 1's to accommodate the intercept such that

\deqn{\Large \mathrm{logit}(\rho_{i,t,j}) = \boldsymbol{a w_i}}
}
