% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_occ.R
\name{auto_occ}
\alias{auto_occ}
\title{Fit an autologistic occupancy model}
\usage{
auto_occ(
  formula,
  y,
  det_covs = NULL,
  occ_covs = NULL,
  method = "BFGS",
  level = 0.95,
  ...
)
}
\arguments{
\item{formula}{Double right-hand side formula describing covariates of
detection (rho) and occupancy (psi) in that order.}

\item{y}{A three-dimensional array of species detections. The first dimension
is sites, the second dimension denotes primary sampling periods, and the third
dimension denotes the secondary sampling periods within each primary sampling
period. If the species was detected on a given survey, that element would receive
a 1, otherwise it is 0. If sampling did not occur for a given sampling period, those
elements should be NA.}

\item{det_covs}{A named list of detection covariates. See details for additional
information on how to format this list.}

\item{occ_covs}{Either a data.frame or named list of occupancy covariates. If
there are temporally varying covariates, use a list. If there are no temporally
varying covariates, use a data.frame. See details for additional infomration
for how to format this.}

\item{method}{The optimization method used by \code{\link[stats]{optim}}.}

\item{level}{The confidence interval size to be calculated for model parameters.
Defaults to 0.95.}

\item{...}{additional arguments. Not used yet.}
}
\description{
Given repeat sampling across multiple primary sampling periods,
this model accounts for first-order temporal autocorrelation in species
occupancy patterns at a site with an additional parameter added to the
latent state of the model (\code{theta}). See Details for additional information.
}
\details{
The autologistic occupancy model can be seen as a very simplified
version of a dynamic occupancy model that makes inference on species
patterns of occupancy from one time period to the next such that species
presence at a site during one time step may modify the probability a species
occupies that same site in the following time step. This is done through the
inclusion of a single logit-scale autoregressive term into the model: \eqn{\theta}.
For \eqn{i} in \eqn{1 \dots I} sites and \eqn{t} in \eqn{1 \dots T} primary sampling periods, let
\eqn{z_{i,t}} be a species latent occupancy state. For T=1, we do not
know if a species occupied a site in a previous time step
(as sampling has not occurred). THus, given a vector of parameters (\eqn{\boldsymbol{\beta}};
 i.e., the occupancy intercept and slope terms) and a matrix of
 covariates whose leading column is a vector of 1's (\eqn{\boldsymbol{X}}),
 the probability of occupancy during the first time step is the logit-linear predictor
at the first time step is:

\deqn{\Large\mathrm{logit}(\psi_{i,t=1}) = \boldsymbol{\beta}\boldsymbol{x}_{i}}

where

\deqn{\Large z_{i,t=1}\sim \mathrm{Bernoulli}(\psi_{i,t=1})}

Following the first time-step, we introduce \eqn{\theta} to help account for changes in occupancy
at time t given their presence at the same site at time t-1. Thus, for the remaining
time periods the logit-linear predictor is

\deqn{\Large \mathrm{logit}(\psi{i,t}) = \boldsymbol{\beta}\boldsymbol{x}_{i} + \theta \times z_{i,t-1} }


where

\deqn{\Large z_{i,t}\sim \mathrm{Bernoulli}(\psi_{i,t}), t>1}

For the data model, let \eqn{\boldsymbol{Y}} be a site by primary sampling period by observation period
array (i.e., a three dimensional array). Thus, for \eqn{i} in \eqn{1 \dots I} sites, \eqn{t} in \eqn{1 \dots T} primary sampling periods and
\eqn{j} in \eqn{1 \dots J} observation events within a given time period (i.e., secondary samples within a primary sampling period), the scalar
\eqn{y_{i,t,j}} can the value of 1 if the species was detected, 0 if it was not detected but sampling occurred, and NA
if sampling did not occur. Therefore, you can consider the vector \eqn{y_{i,t,1:J}} the detection history for a species at site \eqn{i} and time
\eqn{t}. When creating this array for \code{auto_occ}, it is important that \eqn{J} equal the max number of
observation events that happened across all time periods. Thus, if there is variation in the amount of sampling
events across time periods you can pad the time periods with less data with NA values.

Given a vector of parameters (\eqn{\boldsymbol{\alpha}};
 i.e., the detection intercept and slope terms) and a matrix of
 covariates whose leading column is a vector of 1's (\eqn{\boldsymbol{W}}),
 the logit-linear predictor for the data model is

\deqn{\Large\mathrm{logit}(\rho_{i,t,j}) = \boldsymbol{\alpha}\boldsymbol{w}_{i}}

where

\deqn{\Large y_{i,t,j}\sim \mathrm{Bernoulli}(\rho_{i,t,j} \times z_{i,t})}

Note that in this example detection probability only varied by sites, but this
level of the model can incorporate covariates that vary be site, time period, or
observation.

Just like with a dynamic occupancy model, estimates of site occupancy
is a derived parameter, and is most easily handled with using
\code{predict()}. Nevertheless, calculating expected occupancy
can be done with the following equation:

\deqn{\Large
 \frac{
   \mathrm{ilogit}(\boldsymbol{\beta}\boldsymbol{x}_{i})
 }{
 \mathrm{ilogit}(\boldsymbol{\beta}\boldsymbol{x}_{i}) + (1 - \mathrm{ilogit}(\boldsymbol{\beta}\boldsymbol{x}_{i} + \theta))
 }
}

Keeping in mind that your covariate structure may vary (here there
is simply variation by sites). Again, \code{predict()} can easily
handle this for you.
}
