% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,auto_occ_fit-method}
\alias{predict,auto_occ_fit-method}
\title{Makes predictions from an autologistic occupancy model}
\usage{
\S4method{predict}{auto_occ_fit}(object, type, newdata = NULL, backTransform = TRUE, level = 0.95)
}
\arguments{
\item{object}{Object of class inheriting from \code{"auto_occ_fit"}.}

\item{type}{Either \code{"psi"} for occupancy or \code{"rho"} for detection.
See details for how expected occupancy is derived from this model object.}

\item{newdata}{A data.frame of covariates to make predictions from.
See details for more information.}

\item{backTransform}{Should predictions be converted back to the probability
scale. Defaults to \code{TRUE}.}

\item{level}{Tolerance / confidence level for predictions. Defaults to \code{0.95}.}
}
\description{
Predicted values based on a \code{auto_occ_fit} model object.
}
